BREAKING CHANGE: Change disk_space return value to u64 (#153)
See issue #152 for more information.

+++ src/freebsd/mod.rs
@@ -290,7 +290,7 @@ impl GeneralReadout for FreeBSDGeneralReadout {
         Err(ReadoutError::MetricNotAvailable)
     }
 
-    fn disk_space(&self) -> Result<(u128, u128), ReadoutError> {
+    fn disk_space(&self) -> Result<(u64, u64), ReadoutError> {
         shared::disk_space(String::from("/"))
     }
